feat: update the structure
