--- packages/ui/src/components/branch-selector/index.ts
@@ -1,4 +1,0 @@
-import { BranchSelector } from './branch-selector'
-import { BranchSelectorBranchProps, BranchSelectorItem } from './types'
-
-export { BranchSelector, type BranchSelectorBranchProps, type BranchSelectorItem }
