+++ packages/ui/src/components/index.ts
@@ -23,7 +23,6 @@ export * from './theme-selector'
 export * from './theme-selector/types'
 export * from './no-data'
 export * from './tabs'
-export * from './branch-selector'
 export * from './command'
 export * from './search-files'
 export * from './skeleton'
