+++ packages/ui/src/views/repo/repo-summary/repo-summary.tsx
@@ -1,152 +1,220 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
 import { useNavigate } from 'react-router-dom'
 
 import {
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
+  Button,
   ButtonGroup,
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
   Icon,
+  ListActions,
+  NoData,
+  SearchFiles,
+  SkeletonList,
   Spacer,
   StackedList,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
   Text
 } from '@/components'
-import { CommitCopyActions } from '@/components/commit-copy-actions'
-import { getInitials } from '@/utils/utils'
+import { BranchSelectorListProps } from '@views/repo'
+import { BranchSelector } from '@views/repo/components'
 
-export enum SummaryItemType {
-  Folder = 0,
-  File = 1
+import { SandboxLayout } from '../../index'
+import Summary from './components/summary'
+import SummaryPanel from './components/summary-panel'
+import { FileProps } from './repo-summary.types'
+
+interface RepoSummaryViewProps {
+  loading: boolean
+  selectedBranch: string
+  branchList: BranchSelectorListProps[]
+  tagList: BranchSelectorListProps[]
+  selectBranch: (branch: string) => void
+  filesList: string[]
+  navigateToFile: (path: string) => void
+  repository:
+    | {
+        git_ssh_url?: string
+        git_url?: string
+        description?: string
+      }
+    | undefined
+  handleCreateToken: () => void
+  repoEntryPathToFileTypeMap: Map<string, FileProps>
+  files: FileProps[]
+  decodedReadmeContent: string
+  summaryDetails: {
+    default_branch_commit_count?: number
+    branch_count?: number
+    tag_count?: number
+    pull_req_summary?: { open_count: number } | undefined
+  }
+  spaceId: string
+  repoId: string
+  gitRef?: string
+  latestCommitInfo?: {
+    userName: string
+    message: string
+    timestamp: string
+    sha: string | null
+  }
 }
 
-interface UserProps {
-  name: string
-  avatarUrl?: string
-}
-
-export interface FileProps {
-  id: string
-  type: SummaryItemType
-  name: string
-  lastCommitMessage: string
-  timestamp: string
-  user?: UserProps
-  sha?: string
-  path: string
-}
-
-interface PageProps {
-  latestFile: Pick<FileProps, 'user' | 'lastCommitMessage' | 'timestamp' | 'sha'>
-  files?: FileProps[]
-}
-
-export const TopTitle = ({ file }: { file: Pick<FileProps, 'user' | 'lastCommitMessage' | 'timestamp' | 'sha'> }) => {
-  const { user, lastCommitMessage } = file
-
-  return (
-    <ButtonGroup.Root verticalAlign="center" spacing="2">
-      <Avatar size="6">
-        <AvatarImage src={user?.avatarUrl} />
-        <AvatarFallback>
-          <Text size={0} color="tertiaryBackground">
-            {getInitials(user?.name || '')}
-          </Text>
-        </AvatarFallback>
-      </Avatar>
-      <Text size={2} weight="normal" color="tertiaryBackground" wrap="nowrap">
-        {user?.name}
-      </Text>
-      <Text size={2} weight="normal" color="primary" className="line-clamp-1">
-        {lastCommitMessage}
-      </Text>
-    </ButtonGroup.Root>
-  )
-}
-
-export const TopDetails = ({ file }: { file: Pick<FileProps, 'user' | 'lastCommitMessage' | 'timestamp' | 'sha'> }) => {
-  const { sha, timestamp } = file
-  return (
-    <ButtonGroup.Root verticalAlign="center" spacing="2">
-      <CommitCopyActions sha={sha || ''} />
-      <Text size={2} weight="normal" color="tertiaryBackground">
-        |
-      </Text>
-      <Text size={2} weight="normal" color="tertiaryBackground">
-        {timestamp}
-      </Text>
-    </ButtonGroup.Root>
-  )
-}
-
-export const Summary = ({ latestFile, files }: PageProps) => {
+export function RepoSummaryView({
+  loading,
+  selectedBranch,
+  branchList,
+  tagList,
+  selectBranch,
+  filesList,
+  navigateToFile,
+  repository,
+  handleCreateToken,
+  repoEntryPathToFileTypeMap,
+  files,
+  decodedReadmeContent,
+  summaryDetails: { default_branch_commit_count = 0, branch_count = 0, tag_count = 0, pull_req_summary },
+  spaceId,
+  repoId,
+  gitRef,
+  latestCommitInfo
+}: RepoSummaryViewProps) {
   const navigate = useNavigate()
 
+  const renderListContent = () => {
+    if (loading) return <SkeletonList />
+
+    if (repoEntryPathToFileTypeMap.size) {
+      return (
+        <Summary
+          latestFile={{
+            user: { name: latestCommitInfo?.userName || '' },
+            lastCommitMessage: latestCommitInfo?.message || '',
+            timestamp: latestCommitInfo?.timestamp || '',
+            sha: latestCommitInfo?.sha || ''
+          }}
+          files={files}
+        />
+      )
+    }
+
+    return (
+      <NoData
+        insideTabView
+        iconName="no-data-folder"
+        title="No files yet"
+        description={['There are no files in this repository yet.', 'Create new or import an existing file.']}
+        primaryButton={{ label: 'Create file' }}
+        secondaryButton={{ label: 'Import file' }}
+      />
+    )
+  }
+
   return (
-    <>
-      <StackedList.Root>
-        <StackedList.Item disableHover isHeader className="px-3 py-2.5">
-          {latestFile ? (
-            <>
-              <StackedList.Field title={<TopTitle file={latestFile} />} />
-              <StackedList.Field right title={<TopDetails file={latestFile} />} />
-            </>
-          ) : (
-            <Text>No files available</Text>
-          )}
-        </StackedList.Item>
-      </StackedList.Root>
-      <Spacer size={5} />
-      <Table variant="asStackedList">
-        <TableHeader>
-          <TableRow>
-            <TableHead>Name</TableHead>
-            <TableHead>Last commit message</TableHead>
-            <TableHead className="text-right">Date</TableHead>
-          </TableRow>
-        </TableHeader>
-        {files && files.length > 0 ? (
-          <TableBody>
-            {files.map(file => (
-              <TableRow key={file.id} onClick={() => navigate(file.path)}>
-                <TableCell>
-                  <ButtonGroup.Root
-                    direction="horizontal"
-                    verticalAlign="center"
-                    spacing="1.5"
-                    className="cursor-pointer"
-                  >
-                    {file.type === SummaryItemType.File ? (
-                      <Icon name="file" size={14} className="text-tertiary-background" />
-                    ) : (
-                      <Icon name="folder" size={14} className="text-tertiary-background" />
-                    )}
-                    <Text truncate color="primary">
-                      {file.name}
-                    </Text>
-                  </ButtonGroup.Root>
-                </TableCell>
-                <TableCell>
-                  <Text color="tertiaryBackground" className="line-clamp-1">
-                    {file.lastCommitMessage}
-                  </Text>
-                </TableCell>
-                <TableCell className="text-right">
-                  <Text color="tertiaryBackground" wrap="nowrap">
-                    {file.timestamp}
-                  </Text>
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        ) : (
-          <Text>No files available</Text>
-        )}
-      </Table>
-    </>
+    <SandboxLayout.Main hasLeftPanel hasHeader hasSubHeader>
+      <SandboxLayout.Columns columnWidths="1fr 220px">
+        <SandboxLayout.Column>
+          <SandboxLayout.Content>
+            <ListActions.Root>
+              <ListActions.Left>
+                <ButtonGroup.Root>
+                  <BranchSelector
+                    name={selectedBranch}
+                    branchList={branchList}
+                    tagList={tagList}
+                    selectBranch={selectBranch}
+                    repoId={repoId}
+                    spaceId={spaceId}
+                  />
+                  <SearchFiles navigateToFile={navigateToFile} filesList={filesList} />
+                </ButtonGroup.Root>
+              </ListActions.Left>
+              <ListActions.Right>
+                <ButtonGroup.Root>
+                  <DropdownMenu>
+                    <DropdownMenuTrigger asChild>
+                      <Button className="gap-x-2" variant="outline">
+                        Add file
+                        <Icon name="chevron-down" size={11} className="chevron-down" />
+                      </Button>
+                    </DropdownMenuTrigger>
+                    <DropdownMenuContent align="start">
+                      <DropdownMenuItem
+                        key={'create-file'}
+                        onClick={() => {
+                          navigate(`/spaces/${spaceId}/repos/${repoId}/code/new/${gitRef || selectedBranch}/~/`)
+                        }}
+                      >
+                        + Create New File
+                      </DropdownMenuItem>
+                    </DropdownMenuContent>
+                  </DropdownMenu>
+                  {/*  
+                    TODO: require moving and preparing a component from views
+                    <CloneRepoDialog
+                      sshUrl={repository?.git_ssh_url ?? 'could not fetch url'}
+                      httpsUrl={repository?.git_url ?? 'could not fetch url'}
+                      handleCreateToken={handleCreateToken}
+                    />
+                   */}
+                </ButtonGroup.Root>
+              </ListActions.Right>
+            </ListActions.Root>
+            <Spacer size={5} />
+            {renderListContent()}
+            <Spacer size={5} />
+            <StackedList.Root>
+              <StackedList.Item isHeader disableHover>
+                <StackedList.Field title={<Text color="tertiaryBackground">README.md</Text>} />
+                {/* TODO: add component and file editing logic */}
+                <StackedList.Field right />
+              </StackedList.Item>
+              <StackedList.Item disableHover>
+                {/*
+                  TODO: require moving and preparing a component from views
+                  <MarkdownViewer source={decodedReadmeContent || ''} />
+                */}
+              </StackedList.Item>
+            </StackedList.Root>
+          </SandboxLayout.Content>
+        </SandboxLayout.Column>
+        <SandboxLayout.Column>
+          <SandboxLayout.Content className="pl-0">
+            <SummaryPanel
+              title="Summary"
+              details={[
+                {
+                  id: '0',
+                  name: 'Commits',
+                  count: default_branch_commit_count,
+                  iconName: 'tube-sign'
+                },
+                {
+                  id: '1',
+                  name: 'Branches',
+                  count: branch_count,
+                  iconName: 'branch'
+                },
+                {
+                  id: '2',
+                  name: 'Tags',
+                  count: tag_count,
+                  iconName: 'tag'
+                },
+                {
+                  id: '3',
+                  name: 'Open pull requests',
+                  count: pull_req_summary?.open_count || 0,
+                  iconName: 'open-pr'
+                }
+              ]}
+              description={repository?.description}
+            />
+          </SandboxLayout.Content>
+        </SandboxLayout.Column>
+      </SandboxLayout.Columns>
+    </SandboxLayout.Main>
   )
 }
